Load styles/scripts only when shortcodes are used
Conditionally check if the post content contains any of the shortcodes
in which case load enqueue the scripts/styles

A much much better way could have been to keep a flag that defaults to
0. Whenever any of the shortcode callback functions is called, it sets
the flag to 1. Then, in the wp_footer, we can decide whether to load
the scripts or not.

The reason why that method is not used is to avoid loading css in body.

This method could potentially be very slow if the post_content is too
large.